requirements.txt
Django==1.8.2
Markdown==2.6.1
PyJWT==1.1.0
PyYAML==3.11
colorama==0.3.3
dj-database-url==0.3.0
dj-static==0.0.6
django-activity-stream==0.5.1
django-braces==1.8.0
django-cors-headers==1.1.0
django-filter==0.9.2
django-jsonfield==0.9.13
django-oauth-toolkit==0.8.1
django-rest-swagger==0.2.9
djangorestframework==3.1.2
djangorestframework-jwt==1.5.0
djoser==0.3.0
fake-factory==0.5.1
gunicorn==19.3.0
ipdb==0.8
ipython==3.1.0
mando==0.3.2
oauthlib==0.7.2
psycopg2==2.6
radon==1.2.1
six==1.9.0
static3==0.6.1
